Improved Abstraction Layer Memory Usage
The abstraction layer now stores the co-ordinates using chars (signed 8 bits). This decreased memory usage by about 190 bytes.
The down side of this is that existing strength functions must now convert this to a float and divide by 20 to work properly.

+++ more animations/halloween.h
@@ -70,7 +70,7 @@ void spookyBreathe(){
 }
 //profile 3
 float wingsFunction(point pt, float t){
-  float phi = atan(pt.y/abs(pt.x))*12;
+  float phi = atan((float)pt.y/abs(pt.x))*12;
   t = 12*PI/8*sin(2*PI*t);
   return 1/((phi+t)*(phi+t)+1);
 }
